pyproject.toml
[tool.poetry]
name = "vending-machine"
version = "0.1.0"
description = ""
authors = ["sudo-lucifer <krittin.niu@student.mahidol.edu>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.5"
dataclasses-json = "^0.5.7"
djangorestframework-dataclasses = "^1.2.0"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
model-bakery = "^1.9.0"
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]
pre-commit = "~=2.21.0"
black = "^22.12.0"

[tool.black]
line-length = 120
target-version = ['py39']
skip-string-normalization = true
extend-exclude = '''\/migrations\/'''

[tool.isort]
profile = "black"
line_length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
